Rename mflux_ai.mflux_ai to mflux_ai.core
